feat: queue refactoring
