+++ .env.example
@@ -1,5 +1,5 @@
-TRANSACTIONS_CLOSE_PAYMENT_RETRY_QUEUE_NAME=transactionclosuresentevents
-TRANSACTIONS_EXPIRATION_QUEUE_NAME=pagopa-ecommerce-expired-events-queue
+TRANSACTIONS_CLOSE_PAYMENT_RETRY_QUEUE_NAME=pagopa-ecommerce-transactions-close-payment-retry-queue
+TRANSACTIONS_EXPIRATION_QUEUE_NAME=pagopa-ecommerce-transactions-expiration-queue
 TRANSACTIONS_REFUND_RETRY_QUEUE_NAME=pagopa-ecommerce-transactions-refund-retry-queue
 TRANSACTIONS_REFUND_QUEUE_NAME=pagopa-ecommerce-transactions-refund-queue
 QUEUE_CONNECTION_STRING=DefaultEndpointsProtocol=http;AccountName=devstoreaccount1;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;QueueEndpoint=http://storage:10001/devstoreaccount1
