+++ .env.intelliJ.example
@@ -1,1 +1,1 @@
-TRANSACTIONS_CLOSE_PAYMENT_RETRY_QUEUE_NAME=transactionclosuresentevents;TRANSACTIONS_EXPIRATION_QUEUE_NAME=pagopa-ecommerce-expired-events-queue;TRANSACTIONS_REFUND_RETRY_QUEUE_NAME=pagopa-ecommerce-transactions-refund-retry-queue;TRANSACTIONS_REFUND_QUEUE_NAME=pagopa-ecommerce-transactions-refund-queue;QUEUE_CONNECTION_STRING=DefaultEndpointsProtocol=http\;AccountName=devstoreaccount1\;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==\;QueueEndpoint=http://storage:10001/devstoreaccount1;ECOMMERCE_STORAGE_QUEUE_ACCOUNT_NAME=devstoreaccount1;ECOMMERCE_STORAGE_QUEUE_KEY=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;ECOMMERCE_STORAGE_QUEUE_ENDPOINT=http://127.0.0.1:10001/devstoreaccount1;AUTH_REQUESTED_TIMEOUT_SECONDS=100;PAYMENT_TRANSACTIONS_GATEWAY_URI="http://127.0.0.1:8000";PAYMENT_TRANSACTIONS_GATEWAY_READ_TIMEOUT=10000;PAYMENT_TRANSACTIONS_GATEWAY_CONNECTION_TIMEOUT=10000;NODO_CONNECTION_STRING="{\"idPSP\":\"idPsp\",\"idChannel\":\"idChannel\",\"idBrokerPSP\":\"idBrokerPsp\",\"password\":\"password\"}";NODO_URI="http://127.0.0.1:3002/nodo-per-pm/v2";NODO_READ_TIMEOUT=10000;NODO_CONNECTION_TIMEOUT=10000;MONGO_HOST=localhost;MONGO_PORT=27017;MONGO_USERNAME=admin;MONGO_PASSWORD=password;REFUND_RETRY_EVENT_BASE_INTERVAL_SECONDS=60;REFUND_RETRY_EVENT_MAX_ATTEMPTS=3;CLOSE_PAYMENT_RETRY_EVENT_BASE_INTERVAL_SECONDS=60;CLOSE_PAYMENT_RETRY_EVENT_MAX_ATTEMPTS=3
+TRANSACTIONS_CLOSE_PAYMENT_RETRY_QUEUE_NAME=pagopa-ecommerce-transactions-close-payment-retry-queue;TRANSACTIONS_EXPIRATION_QUEUE_NAME=pagopa-ecommerce-transactions-expiration-queue;TRANSACTIONS_REFUND_RETRY_QUEUE_NAME=pagopa-ecommerce-transactions-refund-retry-queue;TRANSACTIONS_REFUND_QUEUE_NAME=pagopa-ecommerce-transactions-refund-queue;QUEUE_CONNECTION_STRING=DefaultEndpointsProtocol=http\;AccountName=devstoreaccount1\;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==\;QueueEndpoint=http://storage:10001/devstoreaccount1;ECOMMERCE_STORAGE_QUEUE_ACCOUNT_NAME=devstoreaccount1;ECOMMERCE_STORAGE_QUEUE_KEY=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;ECOMMERCE_STORAGE_QUEUE_ENDPOINT=http://127.0.0.1:10001/devstoreaccount1;AUTH_REQUESTED_TIMEOUT_SECONDS=100;PAYMENT_TRANSACTIONS_GATEWAY_URI="http://127.0.0.1:8000";PAYMENT_TRANSACTIONS_GATEWAY_READ_TIMEOUT=10000;PAYMENT_TRANSACTIONS_GATEWAY_CONNECTION_TIMEOUT=10000;NODO_CONNECTION_STRING="{\"idPSP\":\"idPsp\",\"idChannel\":\"idChannel\",\"idBrokerPSP\":\"idBrokerPsp\",\"password\":\"password\"}";NODO_URI="http://127.0.0.1:3002/nodo-per-pm/v2";NODO_READ_TIMEOUT=10000;NODO_CONNECTION_TIMEOUT=10000;MONGO_HOST=localhost;MONGO_PORT=27017;MONGO_USERNAME=admin;MONGO_PASSWORD=password;REFUND_RETRY_EVENT_BASE_INTERVAL_SECONDS=60;REFUND_RETRY_EVENT_MAX_ATTEMPTS=3;CLOSE_PAYMENT_RETRY_EVENT_BASE_INTERVAL_SECONDS=60;CLOSE_PAYMENT_RETRY_EVENT_MAX_ATTEMPTS=3
