+++ src/main/kotlin/it/pagopa/ecommerce/scheduler/config/QueuesConsumerConfig.kt
@@ -10,60 +10,45 @@ import org.springframework.integration.annotation.Poller
 
 @Configuration
 class QueuesConsumerConfig {
-    @Bean
-    @InboundChannelAdapter(
-        channel = "transactionactivatedchannel", poller = [Poller(fixedDelay = "1000")]
-    )
-    fun storageQueueActivatedMessageSource(
-        storageQueueTemplate: StorageQueueTemplate,
-        @Value("\${azurestorage.queues.transactionactivatedevents.name}")
-        queueNameActivatedEvents: String
-    ): StorageQueueMessageSource {
-        return StorageQueueMessageSource(queueNameActivatedEvents, storageQueueTemplate)
-    }
 
-    @Bean
-    @InboundChannelAdapter(
-        channel = "transactionclosureschannel", poller = [Poller(fixedDelay = "1000")]
-    )
-    fun storageQueueClosuresMessageSource(
-        storageQueueTemplate: StorageQueueTemplate,
-        @Value("\${azurestorage.queues.transactionclosepaymentretry.name}")
-        queueNameClosureEvents: String
-    ): StorageQueueMessageSource {
-        return StorageQueueMessageSource(queueNameClosureEvents, storageQueueTemplate)
-    }
+  @Bean
+  @InboundChannelAdapter(
+    channel = "transactionclosureschannel", poller = [Poller(fixedDelay = "1000")])
+  fun storageQueueClosuresMessageSource(
+    storageQueueTemplate: StorageQueueTemplate,
+    @Value("\${azurestorage.queues.transactionclosepaymentretry.name}")
+    queueNameClosureEvents: String
+  ): StorageQueueMessageSource {
+    return StorageQueueMessageSource(queueNameClosureEvents, storageQueueTemplate)
+  }
 
-    @Bean
-    @InboundChannelAdapter(
-        channel = "transactionexpiredchannel", poller = [Poller(fixedDelay = "1000")]
-    )
-    fun storageQueueExpirationsMessageSource(
-        storageQueueTemplate: StorageQueueTemplate,
-        @Value("\${azurestorage.queues.transactionexpiredevents.name}") queueNameClosureEvents: String
-    ): StorageQueueMessageSource {
-        return StorageQueueMessageSource(queueNameClosureEvents, storageQueueTemplate)
-    }
+  @Bean
+  @InboundChannelAdapter(
+    channel = "transactionexpiredchannel", poller = [Poller(fixedDelay = "1000")])
+  fun storageQueueExpirationsMessageSource(
+    storageQueueTemplate: StorageQueueTemplate,
+    @Value("\${azurestorage.queues.transactionexpiration.name}") queueNameClosureEvents: String
+  ): StorageQueueMessageSource {
+    return StorageQueueMessageSource(queueNameClosureEvents, storageQueueTemplate)
+  }
 
-    @Bean
-    @InboundChannelAdapter(
-        channel = "transactionrefundretrychannel", poller = [Poller(fixedDelay = "1000")]
-    )
-    fun storageQueueRefundRetryMessageSource(
-        storageQueueTemplate: StorageQueueTemplate,
-        @Value("\${azurestorage.queues.transactionrefundretry.name}") queueNameRefundRetryEvents: String
-    ): StorageQueueMessageSource {
-        return StorageQueueMessageSource(queueNameRefundRetryEvents, storageQueueTemplate)
-    }
+  @Bean
+  @InboundChannelAdapter(
+    channel = "transactionrefundretrychannel", poller = [Poller(fixedDelay = "1000")])
+  fun storageQueueRefundRetryMessageSource(
+    storageQueueTemplate: StorageQueueTemplate,
+    @Value("\${azurestorage.queues.transactionrefundretry.name}") queueNameRefundRetryEvents: String
+  ): StorageQueueMessageSource {
+    return StorageQueueMessageSource(queueNameRefundRetryEvents, storageQueueTemplate)
+  }
 
-    @Bean
-    @InboundChannelAdapter(
-        channel = "transactionsrefundchannel", poller = [Poller(fixedDelay = "1000")]
-    )
-    fun storageQueueRefundMessageSource(
-        storageQueueTemplate: StorageQueueTemplate,
-        @Value("\${azurestorage.queues.transactionsrefund.name}") queueNameRefundRetryEvents: String
-    ): StorageQueueMessageSource {
-        return StorageQueueMessageSource(queueNameRefundRetryEvents, storageQueueTemplate)
-    }
+  @Bean
+  @InboundChannelAdapter(
+    channel = "transactionsrefundchannel", poller = [Poller(fixedDelay = "1000")])
+  fun storageQueueRefundMessageSource(
+    storageQueueTemplate: StorageQueueTemplate,
+    @Value("\${azurestorage.queues.transactionsrefund.name}") queueNameRefundRetryEvents: String
+  ): StorageQueueMessageSource {
+    return StorageQueueMessageSource(queueNameRefundRetryEvents, storageQueueTemplate)
+  }
 }
