+++ src/main/kotlin/it/pagopa/ecommerce/scheduler/queues/TransactionExpirationQueueConsumer.kt
@@ -3,18 +3,13 @@ package it.pagopa.ecommerce.scheduler.queues
 import com.azure.core.util.BinaryData
 import com.azure.spring.messaging.AzureHeaders
 import com.azure.spring.messaging.checkpoint.Checkpointer
-import it.pagopa.ecommerce.commons.documents.v1.TransactionExpiredEvent
-import it.pagopa.ecommerce.commons.documents.v1.TransactionRefundRetriedEvent
-import it.pagopa.ecommerce.commons.documents.v1.TransactionRefundedData
-import it.pagopa.ecommerce.commons.domain.v1.EmptyTransaction
-import it.pagopa.ecommerce.commons.domain.v1.Transaction
-import it.pagopa.ecommerce.commons.domain.v1.pojos.BaseTransaction
-import it.pagopa.ecommerce.commons.domain.v1.pojos.BaseTransactionWithRefundRequested
-import it.pagopa.ecommerce.commons.generated.server.model.TransactionStatusDto
+import it.pagopa.ecommerce.commons.documents.v1.*
+import it.pagopa.ecommerce.commons.utils.v1.TransactionUtils
 import it.pagopa.ecommerce.scheduler.client.PaymentGatewayClient
 import it.pagopa.ecommerce.scheduler.repositories.TransactionsEventStoreRepository
 import it.pagopa.ecommerce.scheduler.repositories.TransactionsViewRepository
 import it.pagopa.ecommerce.scheduler.services.eventretry.RefundRetryService
+import java.util.*
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
@@ -23,35 +18,39 @@ import org.springframework.messaging.handler.annotation.Header
 import org.springframework.messaging.handler.annotation.Payload
 import org.springframework.stereotype.Service
 import reactor.core.publisher.Mono
-import reactor.kotlin.core.publisher.switchIfEmpty
 
 /**
- * Event consumer for transactions to refund. These events are input in the event queue only when a
- * transaction is stuck in an REFUND_REQUESTED state **and** needs to be reverted
+ * Event consumer for events related to transaction activation. This consumer's responsibilities are
+ * to handle expiration of transactions and subsequent refund for transaction stuck in a
+ * pending/transient state.
  */
 @Service
-class TransactionsRefundConsumer(
+class TransactionExpirationQueueConsumer(
   @Autowired private val paymentGatewayClient: PaymentGatewayClient,
   @Autowired private val transactionsEventStoreRepository: TransactionsEventStoreRepository<Any>,
+  @Autowired
+  private val transactionsExpiredEventStoreRepository:
+    TransactionsEventStoreRepository<TransactionExpiredData>,
   @Autowired
   private val transactionsRefundedEventStoreRepository:
     TransactionsEventStoreRepository<TransactionRefundedData>,
   @Autowired private val transactionsViewRepository: TransactionsViewRepository,
+  @Autowired private val transactionUtils: TransactionUtils,
   @Autowired private val refundRetryService: RefundRetryService
 ) {
 
-  var logger: Logger = LoggerFactory.getLogger(TransactionsRefundConsumer::class.java)
+  var logger: Logger = LoggerFactory.getLogger(TransactionExpirationQueueConsumer::class.java)
 
   private fun getTransactionIdFromPayload(data: BinaryData): Mono<String> {
     val idFromActivatedEvent =
-      data.toObjectAsync(TransactionExpiredEvent::class.java).map { it.transactionId }
-    val idFromRefundedEvent =
-      data.toObjectAsync(TransactionRefundRetriedEvent::class.java).map { it.transactionId }
+      data.toObjectAsync(TransactionActivatedEvent::class.java).map { it.transactionId }
+    val idFromClosedEvent =
+      data.toObjectAsync(TransactionClosedEvent::class.java).map { it.transactionId }
 
-    return Mono.firstWithValue(idFromActivatedEvent, idFromRefundedEvent)
+    return Mono.firstWithValue(idFromActivatedEvent, idFromClosedEvent)
   }
 
-  @ServiceActivator(inputChannel = "transactionsrefundchannel", outputChannel = "nullChannel")
+  @ServiceActivator(inputChannel = "transactionexpiredchannel", outputChannel = "nullChannel")
   fun messageReceiver(
     @Payload payload: ByteArray,
     @Header(AzureHeaders.CHECKPOINTER) checkpointer: Checkpointer
@@ -59,24 +58,27 @@ class TransactionsRefundConsumer(
     val checkpoint = checkpointer.success()
 
     val transactionId = getTransactionIdFromPayload(BinaryData.fromBytes(payload))
-
+    val baseTransaction = reduceEvents(transactionId, transactionsEventStoreRepository)
     val refundPipeline =
-      transactionId
-        .flatMapMany { transactionsEventStoreRepository.findByTransactionId(it) }
-        .reduce(EmptyTransaction(), Transaction::applyEvent)
-        .cast(BaseTransaction::class.java)
-        .filter { it.status == TransactionStatusDto.REFUND_REQUESTED }
-        .switchIfEmpty {
-          return@switchIfEmpty transactionId
-            .doOnNext {
-              logger.info("Transaction $it was not previously authorized. No refund needed")
-            }
-            .flatMap { Mono.empty() }
+      baseTransaction
+        .filter { transactionUtils.isTransientStatus(it.status) }
+        .flatMap { tx ->
+          val refundable = isTransactionRefundable(tx)
+          val isTransactionExpired = isTransactionExpired(tx)
+          logger.info(
+            "Transaction ${tx.transactionId.value} in status ${tx.status}, refundable: $refundable, expired: $isTransactionExpired")
+          if (!isTransactionExpired) {
+            updateTransactionToExpired(
+              tx, transactionsExpiredEventStoreRepository, transactionsViewRepository, refundable)
+          } else {
+            Mono.just(tx)
+          }
         }
-        .doOnNext {
-          logger.info("Handling refund request for transaction with id ${it.transactionId.value}")
+        .filter { isTransactionRefundable(it) }
+        .flatMap {
+          updateTransactionToRefundRequested(
+            it, transactionsRefundedEventStoreRepository, transactionsViewRepository)
         }
-        .cast(BaseTransactionWithRefundRequested::class.java)
         .flatMap { tx ->
           refundTransaction(
             tx,
