+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-azurestorage.queues.transactionexpiredevents.name=${TRANSACTIONS_EXPIRATION_QUEUE_NAME}
+azurestorage.queues.transactionexpiration.name=${TRANSACTIONS_EXPIRATION_QUEUE_NAME}
 azurestorage.queues.transactionsrefund.name=${TRANSACTIONS_REFUND_QUEUE_NAME}
 azurestorage.queues.transactionrefundretry.name=${TRANSACTIONS_REFUND_RETRY_QUEUE_NAME}
 azurestorage.queues.transactionclosepaymentretry.name=${TRANSACTIONS_CLOSE_PAYMENT_RETRY_QUEUE_NAME}
