+++ src/test/kotlin/it/pagopa/ecommerce/scheduler/queues/TransactionsRefundEventsConsumerTests.kt
@@ -49,7 +49,7 @@ class TransactionsRefundEventsConsumerTests {
   private val transactionsViewRepository: TransactionsViewRepository = mock()
 
   private val transactionRefundedEventsConsumer =
-    TransactionsRefundConsumer(
+    TransactionsRefundQueueConsumer(
       paymentGatewayClient,
       transactionsEventStoreRepository,
       transactionsRefundedEventStoreRepository,
