+++ src/test/resources/application.test.properties
@@ -1,7 +1,7 @@
 azurestorage.queues.transactionauthrequestedevents.name=transactionauthrequestedtevents
 azurestorage.queues.transactionactivatedevents.name=transactionactivatedtevents
 azurestorage.queues.transactionclosepaymentretry.name=transactionclosepaymentretry
-azurestorage.queues.transactionexpiredevents.name=transactionexpiredevents
+azurestorage.queues.transactionexpiration.name=transactionexpiration
 azurestorage.queues.transactionsrefund.name=transactionrefundevents
 azurestorage.queues.transactionrefundretry.name=transactionrefundretryevents
